# swaggerv2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ecommerce RESTful API",
    "version": "2.0.0",
    "description": "A backend RESTful API for an Ecommerce Store",
    "contact": {
      "name": "Custom Haven",
      "url": "https://github.com/CustomHaven"
    },
    "license": {
      "name": "MIT",
      "url": "https://choosealicense.com/licenses/mit/"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4001/",
      "variables": {}
    },
    {
      "url": "https://rest-api-no-orm.herokuapp.com/",
      "variables": {}
    }
  ],
  "paths": {
...
    "/v2/payment-details": {
      "summary": "Routes dealing with payment details",
      "description": "These are all fake examples of fake credit cards",
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "All payment only for admin",
        "description": "Only for administrator all payment details",
        "operationId": "get_all_payments",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "[{\n  \"id\": 1,\n  \"user_id\": \"user 1\",\n  \"name_on_card\": \"Glen Rey\",\n  \"card_type\": \"mastercard\",\n  \"card_number\": \"5436178723473028\",\n  \"expiry_date\": \"2023-02\",\n  \"cvv\": 761,\n  \"created\": \"07-09-2021\",\n  \"modified\": \"07-09-2021\"\n}, {\n  \"id\": 2,\n  \"user_id\": \"user 2\",\n  \"name_on_card\": \"Nettie Burel\",\n  \"card_type\": \"jcb\",\n  \"card_number\": \"3562467450368308\",\n  \"expiry_date\": \"2024-04\",\n  \"cvv\": 653,\n  \"created\": \"08-09-2021\",\n  \"modified\": \"08-09-2021\"\n}, {\n  \"id\": 3,\n  \"user_id\": \"user 3\",\n  \"name_on_card\": \"Shayla Mustarde\",\n  \"card_type\": \"bankcard\",\n  \"card_number\": \"5602221090906614\",\n  \"expiry_date\": \"2023-02\",\n  \"cvv\": 591,\n  \"created\": \"08-08-2022\",\n  \"modified\": \"08-08-2022\"\n}]"
                  }
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Forbidden\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Payment Details"
        ]
      }
    },
    "/v2/payment-details/publish-key": {
      "summary": "The API publish key for client use",
      "description": "API publish key route for the client to use",
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "The Stripe Publish Key",
        "description": "The API publish key for client use",
        "operationId": "stripe_publish_key",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{\n  \"publishKey\": \"pk_test_RANDOM_STRING\"\n }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Forbidden\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Payment Details"
        ]
      }
    },
    "/v2/payment-details/create-payment-intent": {
      "summary": "The Paying using safe stripe option",
      "description": "This route is paying through stripe",
      "post": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Frontend client pays through stripe interface",
        "description": "Client pays through Stripe using the publish key",
        "operationId": "payment_through_stripe_using_publish_key",
        "requestBody": {
          "description": "Pay using stripe publish key",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Payment-Intent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": {
                      "payment_intent_id": "pi_A_STRING_CONFIRMING_PAYMENT"
                    }
                  }
                }
              }
            },
            "description": "Success"
          },
          "402": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": {
                      "message": "Payment declined"
                    }
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": {
                      "message": "Internal Server Error"
                    }
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Payment Details"
        ]
      }
    },
    "/v2/payment-details/{user_id}": {
      "summary": "The route to make a payment",
      "description": "This route is to make a payment",
      "post": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Make a payment",
        "description": "A customer makes a payment through this route",
        "operationId": "make_a_payment",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "The id of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Sign up a new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Payment-Detail"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{\"id\":4,\"user_id\":\"user 4\", \"name_on_card\":\"Kev Perkins\", \"card_type\":\"maestro\", \"card_number\":\"6759286189048599428\",\"expiry_date\":\"2024-10\", \"cvv\":346,\"created\":\"15-08-2022\",\"modified\":\"15-08-2022\"}"
                  }
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Invalid information\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Payment Details"
        ]
      }
    },
    "/v2/payment-details/{id}": {
      "summary": "Routes for specific payment detail",
      "description": "These routes can only be accessed by a customer and they will only see their own payment details, if they want to manipulate it or not.",
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Show customer their payment details",
        "description": "Show only the customer their own payment detail",
        "operationId": "show_customer_their_payment_detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the payment detail",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{\n  \"id\": 2,\n  \"user_id\": \"user 2\",\n  \"name_on_card\": \"Nettie Burel\",\n  \"card_type\": \"jcb\",\n  \"card_number\": \"3562467450368308\",\n  \"expiry_date\": \"2024-04\",\n  \"cvv\": 653,\n  \"created\": \"08-09-2021\",\n  \"modified\": \"08-09-2021\"\n}"
                  }
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Forbidden\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Payment Details"
        ]
      },
      "put": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Enable customer to update their payment details",
        "description": "Enable the logged in customer to update their own payment detail",
        "operationId": "update_customer_payment_detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the payment detail",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update customer's payment details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Payment-Detail"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{\n  \"id\": 2,\n  \"user_id\": \"user 2\",\n  \"name_on_card\": \"Nettie Burel\",\n  \"card_type\": \"jcb\",\n  \"card_number\": \"3554468495886296\",\n  \"expiry_date\": \"2027-08\",\n  \"cvv\": 206,\n  \"created\": \"08-09-2021\",\n  \"modified\": \"24-08-2022\"\n}"
                  }
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Forbidden\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Payment Details"
        ]
      },
      "delete": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Delete customer's payment details",
        "description": "A customer can delete their own payment details",
        "operationId": "delete_customer_payment_detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the payment detail",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{\n  \"message\": \"Successfully deleted\"\n}"
                  }
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Forbidden\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Payment Details"
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Payment-Detail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "user_id": {
            "type": "string"
          },
          "name_on_card": {
            "type": "string"
          },
          "card_type": {
            "type": "string"
          },
          "card_number": {
            "type": "number"
          },
          "expiry_date": {
            "type": "string"
          },
          "cvv": {
            "type": "number"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Payment-Detail"
        }
      },
      "Payment-Intent": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "payment_method": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Payment-Intent"
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect_sid"
      }
    }
  }
}